sync

+++ tests/tcp_sctp_client.cc
@@ -326,9 +326,9 @@ int main(int ac, char ** av) {
         auto test = config["test"].as<std::string>();
         auto ncon = config["conn"].as<unsigned>();
         auto proto = config["proto"].as<std::string>();
-        auto time = config["time"].as<size_t>();
+        auto time = config["time"].as<unsigned>();
 
-        size_t total_transmission_bytes = static_cast<size_t>(1024*1024*1024)*time/static_cast<size_t>(8);
+        size_t total_transmission_bytes = static_cast<size_t>(1024*1024*1024)*static_cast<size_t>(time)/static_cast<size_t>(8);
         total_transmission_bytes *= 10;
         size_t per_connection_transmission_bytes = total_transmission_bytes/static_cast<size_t>((ncon*smp::count));
         tx_msg_nr = per_connection_transmission_bytes/tx_msg_size + 1;
